ITS ALIIIIIIVE\nUART замість USB CDC

+++ Module_1_7/temp_light_display/STM32CubeIDE/Application/User/Core/inc/sensor.h
@@ -3,6 +3,6 @@
 #ifndef APPLICATION_USER_CORE_INC_SENSOR_H_
 #define APPLICATION_USER_CORE_INC_SENSOR_H_
 
-void get_calculated_sensor_values(uint16_t* raw, uint16_t* output);
+void get_calculated_sensor_values(uint16_t* raw, float* output);
 
 #endif /* APPLICATION_USER_CORE_INC_SENSOR_H_ */
